tweak(adminPanel): RN-1214: Enable expand for sql editor and change icon (#5589)
* add icon

* rearrange components and change expand icon

* Update packages/admin-panel/src/VizBuilderApp/components/ExpandIcon.jsx


* Update packages/admin-panel/src/VizBuilderApp/components/DataLibrary/component/TransformEditor.jsx


* Remove unneeded icon file

---------
