+++ packages/admin-panel/src/VizBuilderApp/components/DataLibrary/component/TransformSelectedOptionWithEditor.jsx
@@ -3,10 +3,14 @@
  * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
  */
 
-import React from 'react';
+import React, { useRef, useState } from 'react';
+import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { SelectedOptionWithJsonEditor } from './SelectedOptionWithJsonEditor';
-import { SelectedOptionWithSqlEditor } from './SelectedOptionWithSqlEditor';
+import { SqlEditor } from '@tupaia/ui-components';
+import { SelectedOptionWithEditor } from './SelectedOptionWithEditor';
+import { JsonEditor } from '../../../../widgets';
+import { TransformModal } from './TransformModal';
+import { TransformEditor } from './TransformEditor';
 
 const getDefaultValueByType = type => {
   switch (type) {
@@ -22,6 +26,25 @@ const getDefaultValueByType = type => {
       return null;
   }
 };
+
+const EditorPanel = styled.div`
+  display: flex;
+  flex: 1;
+  width: 100%;
+
+  > div {
+    width: 100%;
+  }
+
+  .jsoneditor {
+    border: none;
+    cursor: text;
+  }
+
+  .ace_editor {
+    width: 100%;
+  }
+`;
 
 export const TransformSelectedOptionWithEditor = ({
   option,
@@ -58,32 +81,64 @@ export const TransformSelectedOptionWithEditor = ({
     description: option.description || '',
   };
 
-  if (code === 'sql') {
-    return (
-      <SelectedOptionWithSqlEditor
-        option={option}
-        basicOption={basicOption}
-        supportsTitleEditing
-        onRemove={onRemove}
-        currentValue={currentValue}
-        setIsDragDisabled={setIsDragDisabled}
-        onChange={onChange}
-      />
-    );
-  }
+  const editorRef = useRef(null);
+  const [isExpanded, setIsExpanded] = useState(false);
+  const [isEnlarged, setEnlarged] = useState(false);
+  const onOpenModal = () => {
+    setEnlarged(true);
+  };
+
+  const onCloseModal = () => {
+    setEnlarged(false);
+    /**
+     * This is a workaround for the fact that in the jsoneditor library, when mode is `code`, if we make it listen to changes to the `value` and then set the value, it will refocus and reset the state of the editor, which is annoying for the user. Instead we can assume we only need to update the json editor in the panel if it is expanded when the modal closes. If it is not expanded when the modal closes, the action of expanding it will trigger the values to be set anyway
+     */
+    if (isExpanded && editorRef.current) {
+      editorRef.current.set(currentValue);
+    }
+  };
+
+  const onToggleExpanded = () => {
+    const newIsExpanded = !isExpanded;
+    if (!newIsExpanded) {
+      // When collapsing, any invalid state is thrown away. To tell our parents
+      // this we have to trigger an onChange event
+      onChange(option);
+    }
+    setIsExpanded(newIsExpanded);
+  };
+
+  const commonProps = {
+    code,
+    currentValue,
+    onChange,
+    optionMetaData,
+    onInvalidChange,
+    option,
+    basicOption,
+    supportsTitleEditing: true,
+  };
 
   return (
-    <SelectedOptionWithJsonEditor
-      option={option}
-      basicOption={basicOption}
-      supportsTitleEditing
-      onRemove={onRemove}
-      currentValue={currentValue}
-      setIsDragDisabled={setIsDragDisabled}
-      optionMetaData={optionMetaData}
-      onInvalidChange={onInvalidChange}
-      onChange={onChange}
-    />
+    <>
+      {isEnlarged && (
+        <TransformModal onClose={onCloseModal}>
+          <TransformEditor {...commonProps} isExpanded isModal />
+        </TransformModal>
+      )}
+
+      <TransformEditor
+        {...commonProps}
+        setIsDragDisabled={setIsDragDisabled}
+        editorRef={ref => {
+          editorRef.current = ref;
+        }}
+        onRemove={onRemove}
+        isExpanded={isExpanded}
+        onToggleExpanded={onToggleExpanded}
+        onOpenModal={onOpenModal}
+      />
+    </>
   );
 };
