+++ packages/admin-panel/src/VizBuilderApp/components/DataLibrary/component/index.js
@@ -1,4 +1,3 @@
-export { AggregateSelectedOptionWithJsonEditor } from './AggregateSelectedOptionWithJsonEditor';
 export { TransformSelectedOption } from './TransformSelectedOption';
 export { TransformSelectedOptionWithEditor } from './TransformSelectedOptionWithEditor';
 export { DataLibrary } from './DataLibrary';
